decouple osworld from vm backend
