+++ cubes/osworld-cube/tests/test_cube.py
@@ -1,9 +1,9 @@
 import pytest
 
-from cube_vm_backend import LocalQEMUVMBackend
+from osworld_cube.vm_backend import OSWorldQEMUVMBackend
 from osworld_cube.debug import get_debug_task_configs, make_debug_agent
 
-_DEBUG_TASK_CONFIGS = {tc.task_id: tc for tc in get_debug_task_configs(vm_backend=LocalQEMUVMBackend())}
+_DEBUG_TASK_CONFIGS = {tc.task_id: tc for tc in get_debug_task_configs(vm_backend=OSWorldQEMUVMBackend())}
 
 
 def run_debug_episode(task_id: str, max_steps: int = 20) -> dict:
