+++ cubes/osworld-cube/tests/test_run_osworld.py
@@ -15,7 +15,7 @@ from __future__ import annotations
 
 from cube.core import ImageContent, Observation, TextContent
 from cube.task import TaskMetadata
-from cube_vm_backend import LocalQEMUVMBackend
+from osworld_cube.vm_backend import OSWorldQEMUVMBackend
 from osworld_cube.computer import ComputerConfig
 from osworld_cube.task import OSWorldTask
 
@@ -37,7 +37,7 @@ def test_instantiate_and_get_first_obs():
         require_a11y_tree=True,
         observe_after_action=False,
     )
-    vm_backend = LocalQEMUVMBackend()
+    vm_backend = OSWorldQEMUVMBackend()
 
     task = OSWorldTask(metadata=metadata, tool_config=tool_config, vm_backend=vm_backend)
